Explicitly set tile mapping for fully connected layer.
Summary:
The main difference is that we take into account both the number of rows
in the matrix and the number of columns to decide how many vertices to
split each tile into. We also don't require that the number of vertices
per row exactly divides the number of elements in the row.

With these changes the amount of memory goes down slightly
(300580898 -> 300255442) as does the total number of cycles
(268520 -> 267426). The number of cycles spent each fully
connected layer is as follows:

| Size      | Before | After |
| 4096x9216 | 11011  | 10386 |
| 4096x4096 | 5978   | 5562  |
| 1000x4096 | 2394   | 2341  |

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D182