Refactor layer construction in neural net benchmarks
Summary:
It is getting increasingly harder to pass information around when
constructing the layers. This commit refactors the Layer base class to
know more about the network and provide that information to its sub-
classes.

Since the layers are specified before the network, the class is split
into two. The LayerSpec class contains the specification of the layer
independent of the network. The Layer class represents a layer tied to a
specify network.

There is no functionality change.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D244