Add implementation of convolution neural networks that use the new convolution instructions
Summary:
Add an implementation of convolutional layers that split the calculation
by sub-sets of input channels. Splitting the layer this way allows the
use of the new convolutional instructions in the architecture.

Currently, this implementation is used for FP16, 3x3 and 5x5 kernels if
they have stride 1.

To implement this method the tensors passed between layers need to be
arranged in a particular way. This is achieved by adding a method to the
layer class that allows its previous layer to enquire as to how it wants
its input tensor layout to be structure.

This change gives considerable speedup due to the extra grunt of the
convolution instructions. The forward pass of the Alexnet benchmark goes
from ~155us to ~91us.

Reviewers: richard

Reviewed By: richard

Maniphest Tasks: T116

Differential Revision: https://phabricator.sourcevertex.net/D246