Display the number of FLOPs and the ideal number of cycles.
Summary:
Comparing the actual number of compute cycles minus the ideal number gives
a good indication of the overhead during compute. The number of FLOPs
divided by the time can be used to compute a headline number of FLOPS
achieved.

For the convolution layers I haven't included multiplication with zero
padding in the number of FLOPs. As a result the number of FLOPs reported
for Alexnet is smaller than the number of FLOPs in Simon's spreadsheet
(628 million instead of 714 million).

Fixes T126

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T126

Differential Revision: https://phabricator.sourcevertex.net/D250