Reuse graph implementations of similar convolutional neural net layers
Summary:
Change convolutional layers to reuse previously constructed graph layers
if the input dimensions and parameters of the convolution are the same.
To do this each layer works on input, weight and bias tensors that it
owns. To perform the forward pass the actual inputs, weights and biases
are copied into the local tensors. Currently the copying is not profiled
in Poplar so is a hidden cost but is expected to be small.

The layers always perform these copies even if the layer is only used
once. In the Alexnet benchmark this actually results in smaller memory
usage since the live ranges of the tensors are split and more
overlapping can occur.

In the resnet example this change greatly reduces the memory usage
from 559MB to 239MB.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D259