Add residual convolutional neural nets
Summary:
Add residual convolutional layers to the neural net benchmarks. These
are layers that add in the output of a previous layer before the
calculation of the non-linearity. See:

    Deep Residual Learning for Image Recognition
    Kaiming He, Xiangyu Zhang, Shaoqing Ren, Jian Sun
    http://arxiv.org/abs/1512.03385

The above paper describes three schemese for converting the residual
previous layer output to match the output of the current layer. Only the
zero-padding option has been implemented at this point.

Reviewers: richard

Reviewed By: richard

Maniphest Tasks: T114

Differential Revision: https://phabricator.sourcevertex.net/D260