Rewrite convolutional layer to explicitly map vertices to tiles.
Summary:
This rewrite introduces a cost model that is used to decide how to
partition vertices across tiles. Code for estimating performance has
been moved to a shared header for use by the cost model and the
neural net graph code. Previously there were two code paths depending on
the number of input channels per group - these have been unified,
simplifying the code.

Each ConvPartial vertex is now responsible for computing one output row.
It performs a nx1 convolution across all input rows on that tile that
contribute to the output row. Even thought the ConvPartial
vertex has changed, I've tried to keep the cycle estimate comparable to
the previous implementation to make comparisons easier. We should
revist these cycle estimate in the future (I've left a TODO in the
code).

Previously computations involving input block that contained padding
were not handled correctly. The getInputRange() and
getWeightRange() methods are now used to compute the non padding input
elements the must be read and the corresponding weights. Compared with the
old (incorrect) code the number of FLOPs required for Alexnet as reported
by getNumberOfFlops() has increased from  to 1256978176 to 1315248768.

This bug fix means that we are not necessarily comparing like with like.
With this caveat, the performance figures for alexnet are as follows:

Before
------
Memory: 80.6%
Cycles: 144835

After
-----
Memory: 77.6%
Cycles: 140274

Ref T117
Fixes T123

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T117, T123

Differential Revision: https://phabricator.sourcevertex.net/D258