Avoid splitting partials sums over IPUs when targetting multi-IPU systems.
Summary:
This significantly reduces the amount of global exchange required. The
overall cycle count for Alexnet on 2 IPUs goes from 195634 cycles to
157281 cycles. The overall cycle count for Resnet on 2 IPUs goes from
870879 cycles to 681758 cycles.

I've left the tile mapping for single IPU systems unchanged - I tried
reordering the dimensions to match the dimensions of the input tensor but
counter-intuitively this slightly increased the cycle count.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D265