Extract the code to build the forward program into a separate method.
Summary:
Previously the forward() method was responsible both for creating the
shared forward program and for creating the sequence of copies required.
After this change the createFwdProg method is responsible for creating the
forward program and forward() only needs to create tensor to tensor copies.

I've unified the code that creates copies for the first use of a layer
with the logic that creates copies for subsequent uses. This made it
clear there was a bug in the old code, in particular the residual input was
incorrectly copied after the forward program had been executed when the
layer was reused. This has now been fixed.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D273