Explicitly map ConvReduce vertices to tiles.
Summary:
Compute the activations that are mapped to each tile and create
vertices on that tile for reductions that feed into the computation
of those activations. This avoids unnecessary communication of values
after the reduction step.

With this change the number of cycles required for Alexnet goes from
145378 cycles to 144475 cycles.

Depends on D279

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D280