Automatically create neural benchmark report
It turns out to be quite hard to either maintain a spreadsheet template
or cut and paste data out of the generated spreasheets to create the end
of sprint report. It seems much easier to use a python script.

The run_benchmarks.py script still produces the clean spreadsheets of
data but as an option it can create a custom report extracting the
current information of interest for the report. I've moved this to a
separate file so as to not clutter up the main logic of the
run_benchmarks.py script.