Use the convolution instructions for arbitrary convolutions.
Summary:
A n x n convolution is computed as a sum of n^2 1x1 convolutions. The
ConvPartial1x1 takes a vector of weights and, for each weight,
a list of pointers to inputs that should be convolved with the weights.
For each input pointer there is a corresponding InOut pointer that points
to the partial sums that the vertex should update.

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T155

Differential Revision: https://phabricator.sourcevertex.net/D318