Decide the work per worker based on the fields of ConvPartial1x1InOut.
Summary:
The weightReuseCount is now set per worker context. With this change the
host code is responsible for partitioning the work between workers. The
increases the amount of memory required for ConvPartial1x1InOut vertices,
but this seems better than assuming that there is no additional memory
compared to a single vertex running on one worker context.

Ref T170

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T170

Differential Revision: https://phabricator.sourcevertex.net/D335