Don't split weights that are accessed contiguously across tiles.
Summary:
This reduces the number of instructions used to set the receive pointer.
It also indirectly ensures that, for layers where
partialChansPerGroup * inChansPerGroup is even, we avoid sending messages
where the data is not a whole number of words.

An alternative strategy would be to map all the shared weights to a
randomly chosen tile out of the tiles that read those weights. Compared
to this strategy I obtained the following figures for resnet50:

| Category | Spread across tiles | Map to one tile |
| Total memory | 64.9031% | 64.2996% |
| Tile max memory | 197756 | 198550 |
| Num cycles | 636518 | 634895 |

If we aggressively target multicast the numbers are as follows:

| Category | Spread across tiles | Map to one tile |
| Total memory | 62.8296% | 62.5041% |
| Tile max memory | 192176 | 194074 |
| Num cycles | 622317 | 622149 |

On balance spreading the weights across tiles seems a slightly
better strategy:

1. If the tools did a better job of targeting multicast the overall
   cycle count would be almost identical
2. The above cycle counts don't include copying of the weight tensors.
   Mapping the weights to one tile cause an imbalance of work when
   weights are copied, increasing cycle count.
3. The maximum memory usage on a tile determines whether the program
   fits on the device or not, and this is lower if weights are spread
   across tiles.

Depends on D337

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D341