Switch the exchange type to aggressive multicast.
Summary:
Overall this reduces memory usage and the number of cycles.
The results for alexnet50 are as follows:

| Category | Before | After |
| Memory usage | 61.9% | 59.8% |
| Cycles | 636518 | 622317 |

Although overall the cycle count decreases there are some layers
where the cycle count increases. For example the exchange cycle
count for the first max polling layer increases from 1258 cycles
to 1727 cycles. This suggests there would be further room for
improvement if the exchange scheduler was smarter.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D348