Try to rule out candidate partitions before computing the compute cost.
Summary:
If the exchange cost for a partition is greater that the total cost of
the best candidate seen so far then there is no point in computing
the compute cost as this will only make things worse. Explore partitions
with partialChansPerGroupCandidates = 4 first since these are more
likely to yield the best partition.

This reduces the time it takes to run alexnet34b from 4m40s to 3m01s.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D346