Add CPU only functional training for fully connected layer
Summary:
This adds back in the training infrastrure in the Net class and adds the
relevant methods for the FullyConnected layer class to enable a simple
training example.

Currently the approach is to add a vertex per output for the backward
pass and a vertex per neuron for the weight update phase. No attempt to
optimize for the IPU has been made yet and the aim has just been to keep
the code simple. Similarly, there are no tile mappings so training will
not work with the IPUModelEngine.

Reviewers: richard

Reviewed By: richard

Maniphest Tasks: T201

Differential Revision: https://phabricator.sourcevertex.net/D370