Consolidate the knowledge of how activations are partitioned in one place.
Summary:
Previously the logic in mapActivations() had to agree with the logic that
decided how to split the reduce and complete phases reduce and complete phases otherwise you would
get extra communication. Instead of duplicating the logic, move it to the
computeActivationsMapping() method which is queried by both
mapActivations() and the code that constructs the reduce and complete
phases.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D398