Reduce the number of vertices without increasing execution time.
Summary:
The execution time is determined by the worker with the most work. If we
know the amount of work a tile is responsible for we can compute the
maximum amount of work a worker on that tile will be responsible for.
From this we can compute the minimum number workers to use on that tile
such that each worker has no more than this amount of work. This
reduces the number of vertices without penalising execution time.

The effect on resnet50 is as follows:

| Category | Before | After  |
| Memory   | 58.35% | 56.31% |
| Cycles   | 686400 | 686381 |

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D399