Fix usage of padding parameter of conv. layers
Summary:
Previously the padding parameter was not used in determining the
input/output ranges of a convolution for a particular weight. This
happened to work OK for padding of 0 in a 1x1 convolution of padding of
n/2 on each side for a nxn convolution. This change using the padding
parameter in all calculations (though 0 or n/2 seem the likely options).

After this change, the 'padding' value is the amount of padding to add
to *all sides* of the input plane. So a padding value of 1 will make an
X * Y plane into a (X+2) * (Y+2) plane.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D413