Remove dependency between graph construction and Net/Layer class
Summary:
Refactor the neural network code so that the modules that build graphs
for fully connected/convolution/maxpool transforms do not depend on the
Layer class (which is removed completely) or the Net class.

The logic that connects Tensor transforms together into a layered
network is not contained within the Net class.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D488