Use mapActivations() to map activations in the fully connected layer.
Summary:
This is consistent how activations are mapped in the other layers.
mapActivations() tries to spread the activations over fewer tiles so there
is a slight change in memory usage / cycles. For example the number of
cycles for Resnet50 is reduced from 424578 to 424569.

I've removed the FullyConnected vertex type. Partial sums are used for all
the fully connected layers in Alexnet / Resnet34b / Resnet50 and it doesn't
seem worth maintaining extra code to optimize a case that we do not
currently hit.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D487