Target the fully connected backward pass at IPUs.
Summary:
The same layout of weights is used in the forward pass and the backwards
pass. The avoids the need to transpose the weights in memory.

During the backwards pass we must walk down columns of the weight matrix.
This currently uses pointers to each element - in future we could be
vectorized across groups of 4 columns.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D499