Try to flatten the x and y dimensions when performing 1x1 convolutions.
Summary:
If the stride and padding are zero a 1x1 convolution of a x by y input is
identical to a 1x1 convolution on a (x * y) by 1 input. Flattening the x
and y dimensions makes easier to partition the convolutions across the
tiles since there is one fewer dimension to subdivide.

This has the following effect on ResNet 50:

| Category | Before | After  |
| Memory   | 42.7%  | 42.1%  |
| Cycles   | 425175 | 416069 |

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D513