Add new method of targeting the conv inst that needs fewer input channels.
Summary:
Instead of one single pass computing a 1x1 convolution on 16 input
channels it computes a nx1 convolution on 16/n input channels. There is
a slight increase in the loop overhead since n input pointers need to
be loaded each pass.

The treatment of padding in the x axis is the same as with 1x1
convolutions. Padding in the y axis is dealt with in the host program.
In one pass the vertex reads from n input pointers, one for each input row
use in the 4x1 convolution. If some of those rows are padding then the
input pointer is set to a block zeroed memory.

Ref T291

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T291

Differential Revision: https://phabricator.sourcevertex.net/D549