Compute the non linearity derivative from the activations.
Summary:
Previously the code computed the derivative based on the "z" values, i.e.
the output of the convolutions before the non linearity was applied.
After this change the z values are no longer used in the backwards pass
which means they don't need to be retained during the forward pass.

Ref T308

Reviewers: Eddy

Reviewed By: Eddy

Maniphest Tasks: T308

Differential Revision: https://phabricator.sourcevertex.net/D566