Zero partial sums in backwards pass of strided 1x1 convolution.
Summary:
In the backward pass of a strided 1x1 convolutional layer output striding
means that not every output delta was written by the ConvPartial1x1Out
vertex. Zero the partial sums to ensure that every delta is written.
Map the partial sums tensor in zeroPartialSums() to ensure every tensor
element is mapped to a tile.

If a vertex has an output that is not written then the value of the element is
undefined afterwards even it was previously written to another vertex. If this
wasn't the case the tools would never be able to overlap tensors in memory.
Therefore it is invalid to use a vertex with field of type Output to perform a
1x1 strided convolution - use ConvPartialnx1InOutVertex instead and remove
output striding support from ConvPartial1x1Out.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D589