Add tool for running a convolutional layer in isolation.
Summary:
The properties of the layer (height, width, stride, etc.) are specified on
the command line. The tool runs one forward pass and reports the results.

Input data and the model parameters are initialized with random values.
Output data is copied back to the host and compared with results from
a reference model. An error is reported if the results from popnn and the
reference model differ significantly.

This self checking behaviour means that the single_conv_layer executable
can be used to help test popnn. For now I've added one test 3x3
convolutional layer - the set of tests can be expanded later.

Reviewers: davel, godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D655