Use the convolution instruction for the weight update.
Summary:
Currently the convolution instruction is only used for 1x1 stride 1
convolutions. The conv unit is loaded with activations and the inner
loop iterates over deltas channels, accumulating over groups of 16
(x,y) coordinates. Prior to the convolution the activations and deltas
must be rearranged. After the convolution we must rearrange the
weightDeltas to match the layout used for the weights in the forward pass.

Note: the cost estimate in ConvPlan.cpp doesn't include the cost of
rearranging the data - this could be improved in future.

Ref T315.

Reviewers: Eddy

Reviewed By: Eddy

Maniphest Tasks: T315

Differential Revision: https://phabricator.sourcevertex.net/D682