Fix the mapping of activations to tiles when using batching
Summary:
Previously the mapping of activations to tiles was incorrect when
numTiles < batchSize. This is fixed by correctly rounding the tile range
when computing the activation mapping.

Additionally, the mapping of convolution vertices to tiles did not match
the activation mapping properly. The logic in linearizeTileIndices is
fixed in this commit.

Reviewers: godfrey.da.costa

Reviewed By: godfrey.da.costa

Differential Revision: https://phabricator.sourcevertex.net/D857