Add splitting of batch size into batch size groups in the convolution plan
Summary:
The existing tile paritioner for convolution layer divides batches into as many
groups as there are batches. The new partitioner divides the batch into groups
and extends the field by a factor which is equal to the number of batches per
group. The number of batch groups and number of elements in each is determined
by the paritioner based on the overall cost.

Only 1x1 kernel sizes are supported.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D867