Target supertile send / receive in the convolutional layer.
Summary:
Change the tile linearize method such that tiles which compute convolutions
on the same input volume for different output channels are adjacent.
Change the cost model so it prefers to partitioning the convolution
for different the output channels across an even number of tiles.
Change the mapping of the tile index to output channel group range to
avoid the need to communicate different activations to different tiles of
the same supertile in the case where output channels are partitioned
across an even number of tiles.

Together these changes reduce exchange time due to better use of supertile
send / receive. The results for ResNet 50 are as follows:

| Category            | Before | After  |
| Memory              | 45.3%  | 41.7%  |
| Cycles              | 365533 | 350240 |
| Exchange efficiency | 81.9   | 101.4  |

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D855