Add the type of the 'labels' to the template. This helps TensorFlow, which uses signed ints for indexing.
Summary:
In Tensorflow the labels (and other forms of integer indexing) are done using signed int (or signed int64).
This change allows for the CalcLoss vertex to take a signed integer vector for the labels.

Conform to 80 character line limit

Test Plan: run Alexnet benchmark, which instantiates and uses the CalcLoss vertex.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D878