Allow specfication of data type for partials and extend model to add latest convolution unit attributes
Summary:
The current specification in popnn derives type for partials depending on the
use of the convolution unit in Colossus. The new changes to the unit allows
output of partials to be either half or single precision float. The new code models
the coefficient load transfer bandwidth as well as allows specification of data
type for partials.

This change has a dependancy on Differential D913

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D912