Reuse forward convolution sub-graphs when doing backwards pass
Summary:
Change backwards pass in Net.hpp to prefer to do non-fractional
convolutions in the backwards pass where possible. This increased the
chance that the tensor op will reuse a convolution sub-graph from the
forwards pass in the backwards pass thus reducing memory.

This requires splitting out the transformation of weights for the
backwards pass. To avoid memory bloat this transformation is converted
to a tensor operation so will be resused in common layers.

The single conv layer tools is updated to use similar logic to the Net
class so that the tests also test that logic.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D951