Update tile mapping avoid unnecessary post reduce exchange.
Summary:
The old code mapped the reduced tensor based on the partial channel group
but used the output channels per group to decide which elements to compute
on each tile. This resulted in extra exchange after the reduce compute set
(since vertices write to elements of the reduce tensor mapped to different
tiles) and extra exchange before the complete compute set (since the
mapping of the reduced tensor didn't follow the mapping of the output
activations.

Update the code to eliminate exchange after the reduce compute set and
minimize exchange before execution of the complete compute set.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D909