Reduce number of activation views used in weight update of convolution layers using convolution unit
Summary:
The existing implementation uses as many activation views as the product of the
kernel size. For large kernel sizes this adds a significant memory cost in
copying of tensors. The new scheme reduces these views and thus reduces memory
cost of tensor copying.  However, because the new scheme requires padding
of the weight tensors after copying, the number of copy pointers increases.
A heuristic method is employed in the selection between reduced and full views.

The heuristic may need to be revisited if constant tensor methods are improved
in Poplar.

This submission addresses Maniphest task T748

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D966