Fix type of reduced values to data type used by activations
Summary:
In the case when there are no partials for reduction, cast the partial output
to activation data type. This gives some saving in memory. A more efficient
solution would be to produce partials output to have the same data type as
the activation in case there is no reduction needed across tiles.

Add comments

The savings for resnet50 is ~2.3Mb and ~3.1Mb for resnet34. Since the savings are small because of the use of the cast operator, it is moot whether this should go in.

Test Plan: T

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D988