Target weight update using the AOP instruction.
Summary:
Weight update using the AOP instruction replaces the old weight update via
the MAC instruction - I'll remove the unused functions in a follow on
commit.

There is now a separate planning method for weight update that takes the
channel groupings of the activations / deltas / weights as an argument
instead of searching for the best channel grouping. The planning has
been refactored so the vertex is decided in the outer loop and the
number of input channels is decided in the inner loop. This makes it
easier to have different logic in the planning depending on the vertex
type.

Compared to AMP, weight update via AOP appears to reduces the number of
cycles, reduce the number of pointers and reduce the amount of exchange
code.

I've updated the cost model for weight update via AMP to take
into account the cost of reordering. With this change the AOP method is
chosen for some but not all of the ResNet-50 layer. If I patch the code
so that AOP is always used then there is a further performance improvement
suggesting some inaccuracy in the cost model. This needs further
investigation.

Fixes T772

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T772

Differential Revision: https://phabricator.sourcevertex.net/D1051