Introduce optional memory optimisation pass in the conv planner
Summary:
A second memory optimisation pass is added to the convolutional layer planner.
This pass searches for a plan which minimises memory cost at the expense of
cycles cost. The penalty in cycles is bounded and may be specified by a command
line option. The bound is specified as a percentage excess with respect to the
optimum obtained in the first pass. The default value is 0 which imples no
memory optimisation.

Current implementation is only for forward/backward pass and doesn't include
the weight update pass.

single_layer_conv shows reduction in memory but benchmark models show worse
results

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1074