Use reduce() to sum weight gradient deltas.
Summary:
Previously partial sums were reduced and added to the weights in
the same vertex. This caused extra exchange if the reduction happened on
a different tile since the weight had to be sent that tile and back.

Split the reduction and the addition of the weight deltas and
perform the addition of the weight deltas on the tile containing the
weights. We no longer need to send the weight to another tile and it also
simplifies the code as we can reuse the existing reduce() function.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1059