Fix assert in reduce due to dimension mismatch.
Summary:
The reduce function expects the dimensions of the reduced tensor to
match the dimensions of the partials tensor with the first dimension
removed. Previously convolutionWeightUpdateAop() flattened the partials
without flattening the weight deltas causing an assertion in debug builds
of popnn.

Instead of fully flattening the partials only flatten the dimensions we are
reducing across.

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D1088