Avoid unnecessary exchange after cast of partial sums before complete.
Summary:
Previously the castPartials function used the activation mapping to
map cast vertices. The reduced tensor has a different tile mapping to the
activations tensor and so this caused extra exchange after compute.

Add a cast function that accepts tile mapping specified as a list of
regions per tile and teach the reduce function to call the cast function
in the case where there is just one partial.

Fixes T794.

Reviewers: Eddy

Reviewed By: Eddy

Maniphest Tasks: T794

Differential Revision: https://phabricator.sourcevertex.net/D1087