Refactor code that determines the candidate vertex type.
Summary:
Don't consider vertices where the activation type doesn't match the
requested activation type. We consider vertices where the
partial type is larger than the requested partial type, but only if
it is not possible to use the requested partial type.

This also changes getConvVertexTypeCandidates() so that the partial type
for the vertex type candidate that doesn't use the convolution instruction
matches the requested partial type. The ConvPartial vertex supports
half partials so I'm not sure why the use of float partials was hardcoded.

Reviewers: davidn

Reviewed By: davidn

Differential Revision: https://phabricator.sourcevertex.net/D1092