Remove numBatchGroups from Plan and initialize batchesPerGroup in the ctor.
Summary:
Previously batchesPerGroup wasn't set until after the call to
estimateCost. estimateReduceComputeCost asserts that plan.batchesPerGroup
is 1 for a weight update but this failed in debug because it wasn't
initialized yet. The numBatchGroups member is redundant because the
number of groups can be determined from the batch size and the group size.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1099