Assume use of axpy instruction for ConvWeightUpdate cycle estimate.
Summary:
Previously the ConvWeightUpdate vertex supported a reduction of partial
sums. I've removed support for this so the ConvWeightUpdate just adds
the weight deltas to the weights. We can the axpy instruction to
perform this operation at the rate of 4 weights per cycle.

Fixes T815.

Depends on D1094

Reviewers: davel

Reviewed By: davel

Maniphest Tasks: T815

Differential Revision: https://phabricator.sourcevertex.net/D1095