Improve reduce() in the case where the destination tensor is unbalanced.
Summary:
Previously reduce() always mapped the reduction of each element to the
tile containing the element. If the destination tensor is unbalanced we
may be better off if we evenly spread the compute across tiles at the
cost of some post compute exchange to get results to the right location.

Estimate the cost of both methods and pick the best.

This affects ResNet-18 as follows:

| Category | Before | After  |
| Memory   | 183.5  | 183.8  |
| Cycles   | 702309 | 699286 |

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1102