Let the reduce method decide how to efficiently map the reduction to tiles.
Summary:
Depends on D1094, D1102

This affects ResNet-18 training (with AOP forced) as follows:

| Category | Before | After  |
| Memory   | 149.9  | 149.8  |
| Cycles   | 658043 | 661468 |

The reason for the increase in cycles is as follows. Previously the weight
deltas were mapped to where they were computed and now weight deltas are always
mapped to the same tiles as the weights. This means you get exchange at
the end of the reduce compute set instead of at the beginning of compute set
that adds the deltas to the weights. The first bias reduction is overlapped with
the weight update and moving the exchange earlier adds an extra sync.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1104