Choose whether to regroup deltas or weight deltas when using AOP.
Summary:
Model the cost of the rearranging the deltas and the weight deltas
during planning and pick which ever has the lowest cost.

This has no affect on ResNet-50 as it always picks rearrangement of
the deltas. However for a suitably choosen set of layer parameters
it will rearrange the weight deltas instead. For example with the
parameters --height=128 --width=128 --input-channels=16
--output-channels=128 --tiles-per-ipu=128 this change has the
following affect:

| Category | Before | After  |
| Memory   | 30.0   | 28.3   |
| Cycles   | 144134 | 135834 |

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1114