Optimize convolution reduction.
Summary:
We map the reduction based on the tile mapping of the partials not
based on the tile mapping of the reduced result. This minimises the number
of tiles each tile must receive from in the exchange prior to the
reduction. In cases where the number of partials is large we break the
reduction down into multiple stages to further reduce code size.

Fixes T854

Reviewers: davel

Reviewed By: davel

Subscribers: Eddy, matt

Maniphest Tasks: T854

Differential Revision: https://phabricator.sourcevertex.net/D1142