Fix activations mapping when batching is used.
Summary:
df7fb870 introduced a bug where the first batch was spread over the
first numTiles / batchSize tiles but subsequent batches we stored on a
single tile. This was functionally correct but results in terrible
performance. It also increases memory usage as the schedules grow very
large due to the single tile bottleneck.

Reviewers: Eddy

Reviewed By: Eddy

Subscribers: davidn

Differential Revision: https://phabricator.sourcevertex.net/D1208