Remove the batchNum argument to iterateBiasMapping.
Summary:
Previously iterateBiasMapping gave different results depending
on the batchNum argument but this makes no sense - there is just
one set of biases that is shared across the batch.

When iterateBiasMapping was called in convolutionBiasUpdate it used
a batchNum of one but in mapBiases it mapped the biases based on a
batchNum of batchSize - 1. This mismatch caused a small amount of
unnecessary exchange.

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1213