Optimize the backwards weight rearrangement.
Summary:
Perform the rearrangement in two steps. The first step rearranges the
channels groups already on that tile such that the second step is a
rearrangement of larger blocks of weights. This reduces the amount of exchange
code required as we send fewer, larger messages in the second phase.

On ResNet-50 Training (batch 1) this saves 2MB of exchange code and
22 MB of copy pointers. It also reduces the number of cycles from
1418631 to 1406906 since the sort instruction can be used to rearrange
fp16 weights faster than the naive tensor copy code.

Fixes T811.

Reviewers: davel, godfrey.da.costa

Reviewed By: davel, godfrey.da.costa

Subscribers: davel

Maniphest Tasks: T811

Differential Revision: https://phabricator.sourcevertex.net/D1224