Refactor max pooling to use tile mapping introspection to reduce vertex count
Summary:
Change max pooling code now looks at the tile mapping of the output and
and on each tile create a small set of vertices that balance the work to
be done to create those outputs.

For each vertex, the outputs are grouped into vectors of channels for
the same pixel position and corresponding vectors are provided for the
input window that is reduced over.

Reviewers: Eddy

Reviewed By: Eddy

Differential Revision: https://phabricator.sourcevertex.net/D1387