Add a new version of the code that performs the weight update using AMP.
Summary:
This differs from the old code in the following respects:

- The deltas are loaded into the amp unit (the previous code loaded
  activations).
- Where we previously expanded the y axis by creating a view with the
  same row repeated multiple times we now express this as a convolution
  where we are sliding the kernel in the y axis.

The new code doesn't handle striding yet.

This has the following effect on ResNet-50 training (batch 1):

| Category | Before  | After   |
| Memory   | 248.4   | 260.0   |
| Cycles   | 1696261 | 1692883 |

There is a reduction in the amount of exchange code / cycles but this is
counteracted by an increase in copy pointers / cycles:

| Category      | Before  | After   |
| Copy cycles   | 91904   | 114197  |
| Other cycles  | 1604357 | 1578686 |
| Copy pointers | 52.7    | 73.0    |
| Other Memory  | 195.7   | 187     |

Because of the increase in overall memory usage I've put the new
weight update method behind an option (off by default).

Reviewers: davel

Reviewed By: davel

Differential Revision: https://phabricator.sourcevertex.net/D1405