Replace convolution planner object with options structure and optional cache
Summary:
Instead of having a planner object, have an convolution implemtation
options object which is passed in to every operation. Plans are
constructed from the information that is given. If the user wants to
cache the creation of plans to speed up the host program a cache can be
created and passed in with the options structure.

Currently the client application still needs to explicitly create plans
to get required output channel groupings but this will go when the
convolutions are no longer structured that way.

Reviewers: richard

Reviewed By: richard

Differential Revision: https://phabricator.sourcevertex.net/D1562